+++ src/test/java/com/example/payment/transfer/TransferServiceTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.when;
 import com.example.payment.account.AccountRepository;
 import com.example.payment.account.entity.Account;
 import com.example.payment.account.exception.NotExistAccountException;
-import com.example.payment.member.exception.NotMatchPasswordException;
+import com.example.payment.global.exception.NotMatchPasswordException;
 import com.example.payment.transfer.dto.reqeust.TransferRequest;
 import com.example.payment.transfer.exception.NotEnoughWithdrawalMoney;
 import com.example.payment.transferHistory.TransferHistoryRepository;
